requirements.txt
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-oauthlib==1.2.2
python-dotenv==1.1.0
requests==2.32.3
